#1 Restructured main program into its own folder
