+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
     "editor.formatOnSave": true,
     "editor.formatOnType": true,
-    "files.insertFinalNewline": true
+    "files.insertFinalNewline": true,
+    "files.eol": "\n",
 }
