+++ my-cli.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MainService", "MainService\MainService.csproj", "{1AC852B7-532B-461D-A609-C3D47A343584}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1AC852B7-532B-461D-A609-C3D47A343584}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1AC852B7-532B-461D-A609-C3D47A343584}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1AC852B7-532B-461D-A609-C3D47A343584}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1AC852B7-532B-461D-A609-C3D47A343584}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
